Moves code from abstract mapping classes to traits to allow inheritance in mappings
