+++ src/Mapping/FileMapping.php
@@ -4,18 +4,9 @@ declare(strict_types=1);
 
 namespace yjiotpukc\MongoODMFluent\Mapping;
 
-use Doctrine\ODM\MongoDB\Mapping\ClassMetadata;
-use yjiotpukc\MongoODMFluent\Builder\Document\FileBuilder;
-use yjiotpukc\MongoODMFluent\Builder\File;
+use yjiotpukc\MongoODMFluent\Mapping\Traits\FileMappingTrait;
 
 abstract class FileMapping implements Mapping
 {
-    public function load(ClassMetadata $metadata): void
-    {
-        $builder = new FileBuilder();
-        $this->map($builder);
-        $builder->build($metadata);
-    }
-
-    abstract public function map(File $builder): void;
+    use FileMappingTrait;
 }
