+++ src/Mapping/MappedSuperclassMapping.php
@@ -4,19 +4,9 @@ declare(strict_types=1);
 
 namespace yjiotpukc\MongoODMFluent\Mapping;
 
-use Doctrine\ODM\MongoDB\Mapping\ClassMetadata;
-use yjiotpukc\MongoODMFluent\Builder\Document;
-use yjiotpukc\MongoODMFluent\Builder\Document\DocumentBuilder;
+use yjiotpukc\MongoODMFluent\Mapping\Traits\MappedSuperclassMappingTrait;
 
 abstract class MappedSuperclassMapping implements Mapping
 {
-    public function load(ClassMetadata $metadata): void
-    {
-        $builder = new DocumentBuilder();
-        $builder->mappedSuperclass();
-        $this->map($builder);
-        $builder->build($metadata);
-    }
-
-    abstract public function map(Document $builder): void;
+    use MappedSuperclassMappingTrait;
 }
