+++ src/Mapping/Traits/ViewMappingTrait.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace yjiotpukc\MongoODMFluent\Mapping\Traits;
+
+use Doctrine\ODM\MongoDB\Mapping\ClassMetadata;
+use yjiotpukc\MongoODMFluent\Builder\Document\DocumentBuilder;
+use yjiotpukc\MongoODMFluent\Builder\View;
+
+trait ViewMappingTrait
+{
+    public function load(ClassMetadata $metadata): void
+    {
+        $builder = new DocumentBuilder();
+        $this->map($builder);
+        $builder->build($metadata);
+    }
+
+    abstract public function map(View $builder): void;
+}
