+++ src/Mapping/ViewMapping.php
@@ -4,18 +4,9 @@ declare(strict_types=1);
 
 namespace yjiotpukc\MongoODMFluent\Mapping;
 
-use Doctrine\ODM\MongoDB\Mapping\ClassMetadata;
-use yjiotpukc\MongoODMFluent\Builder\Document\DocumentBuilder;
-use yjiotpukc\MongoODMFluent\Builder\View;
+use yjiotpukc\MongoODMFluent\Mapping\Traits\ViewMappingTrait;
 
 abstract class ViewMapping implements Mapping
 {
-    public function load(ClassMetadata $metadata): void
-    {
-        $builder = new DocumentBuilder();
-        $this->map($builder);
-        $builder->build($metadata);
-    }
-
-    abstract public function map(View $builder): void;
+    use ViewMappingTrait;
 }
